+++ addon/components/flexberry-lookup.js
@@ -4,8 +4,8 @@
 
 import { assert, warn, debug } from '@ember/debug';
 import { inject as service } from '@ember/service';
-import { computed, observer } from '@ember/object';
-import { later, run } from '@ember/runloop';
+import { computed, observer, get } from '@ember/object';
+import { later, run, once } from '@ember/runloop';
 import { merge } from '@ember/polyfills';
 import { isNone } from '@ember/utils';
 import { on } from '@ember/object/evented';
@@ -15,9 +15,7 @@ import { translationMacro as t } from 'ember-i18n';
 import { getRelationType } from 'ember-flexberry-data/utils/model-functions';
 import Builder from 'ember-flexberry-data/query/builder';
 import Condition from 'ember-flexberry-data/query/condition';
-import { BasePredicate } from 'ember-flexberry-data/query/predicate';
-import { ComplexPredicate } from 'ember-flexberry-data/query/predicate';
-import { StringPredicate } from 'ember-flexberry-data/query/predicate';
+import { BasePredicate, SimplePredicate, ComplexPredicate, StringPredicate } from 'ember-flexberry-data/query/predicate';
 
 /**
   Lookup component for Semantic UI.
@@ -319,6 +317,24 @@ export default FlexberryBaseComponent.extend({
   maxResults: 10,
 
   /**
+    Flag for remove autocomplete value if it's not in the results (remove when false).
+
+    @property autocompletePersistValue
+    @type Boolean
+    @default false
+  */
+  autocompletePersistValue: false,
+
+  /**
+    Stores autocomplete value, when autocompletePersistValue=true and lookup value clearing.
+
+    @property autocompletePersistValueCache
+    @type String
+    @default undefined
+  */
+  autocompletePersistValueCache: undefined,
+
+  /**
     Limit function on lookup.
     It should not be just a string, it has to be predicate function (otherwise an exception will be thrown).
 
@@ -503,6 +519,15 @@ export default FlexberryBaseComponent.extend({
   autocompleteOrder: null,
 
   /**
+    Projection name for autocomplete query.
+
+    @property autocompleteProjection
+    @type String
+    @default undefined
+  */
+  autocompleteProjection: undefined,
+
+  /**
     Current selected instance of the model.
 
     @property value
@@ -552,11 +577,8 @@ export default FlexberryBaseComponent.extend({
 
     @property displayValue
     @type String
-    @readOnly
   */
-  displayValue: computed('value', function() {
-    return this._buildDisplayValue();
-  }),
+  displayValue: undefined,
 
   /**
     Standard CSS class names to apply to the view's outer element.
@@ -668,6 +690,11 @@ export default FlexberryBaseComponent.extend({
         return;
       }
 
+      if (this.get('autocompletePersistValue')) {
+        this.set('autocompletePersistValueCache', undefined);
+        this.set('displayValue', undefined);
+      }
+
       let remove = this.get('remove');
       if (remove instanceof Function) {
         remove(removeData);
@@ -728,14 +755,20 @@ export default FlexberryBaseComponent.extend({
     this.get('lookupEventsService').on('lookupDialogOnVisible', this, this._setModalIsVisible);
     this.get('lookupEventsService').on('lookupDialogOnHidden', this, this._setModalIsHidden);
 
-    // TODO: This is necessary because of incomprehensible one-way binding on new detail form, perhaps the truth is out there, but I did not find it.
-    this.addObserver('value', this, this._valueObserver);
-    this.addObserver('displayAttributeName', this, this._valueObserver);
-    this.addObserver(`relatedModel.${this.get('relationName')}`, this, this._valueObserver);
+    if (this.get('autocompletePersistValue') && isNone(this.get('value'))) {
+      this.set('autocompletePersistValueCache', this.get('displayValue'));
+    }
+
+    this._valueObserver();
   },
 
   /**
-    It seems to me a mistake here, it should be [willDestroyElement](https://emberjs.com/api/ember/release/classes/Component#event_willDestroyElement).
+    Observe lookup value changes.
+  */
+  valueObserver: observer('value', 'displayAttributeName', function() { once(this, '_valueObserver'); }),
+
+  /**
+    It seems to me a mistake here, it should be [willDestroyElement](http://emberjs.com/api/classes/Ember.Component.html#event_willDestroyElement).
     # WRANING!
 
     @method didDestroyElement
@@ -789,11 +822,6 @@ export default FlexberryBaseComponent.extend({
     this.get('lookupEventsService').off('lookupDialogOnShow', this, this._setModalIsStartToShow);
     this.get('lookupEventsService').off('lookupDialogOnVisible', this, this._setModalIsVisible);
     this.get('lookupEventsService').off('lookupDialogOnHidden', this, this._setModalIsHidden);
-
-    // TODO: This is necessary because of incomprehensible one-way binding on new detail form, perhaps the truth is out there, but I did not find it.
-    this.removeObserver('value', this, this._valueObserver);
-    this.removeObserver('displayAttributeName', this, this._valueObserver);
-    this.removeObserver(`relatedModel.${this.get('relationName')}`, this, this._valueObserver);
   },
 
   /**
@@ -866,6 +894,7 @@ export default FlexberryBaseComponent.extend({
     }
 
     let relationModelName = getRelationType(relatedModel, relationName);
+    let projectionName = this.get('autocompleteProjection');
 
     let displayAttributeName = this.get('displayAttributeName');
     let autocompleteOrder = this.get('autocompleteOrder');
@@ -914,13 +943,30 @@ export default FlexberryBaseComponent.extend({
             return;
           }
 
+          let selectAttributes = displayAttributeName;
+          if (!isNone(projectionName)) {
+            let modelConstructor = store.modelFor(relationModelName);
+            let projection = get(modelConstructor, `projections.${projectionName}`);
+            if (!projection) {
+              throw new Error(`No projection with '${projectionName}' name defined in '${relationModelName}' model.`);
+            }
+
+            let attributes = get(projection, 'attributes');
+            let attributesKeys = Object.keys(attributes);
+            if (attributesKeys.indexOf(displayAttributeName) === -1) {
+              attributesKeys.push(displayAttributeName);
+            }
+
+            selectAttributes = attributesKeys.join(',');
+          }
+
           let builder;
           if (autocompleteOrder) {
             builder = new Builder(store, relationModelName)
-            .select(displayAttributeName).orderBy(`${autocompleteOrder}`);
+            .select(selectAttributes).orderBy(`${autocompleteOrder}`);
           } else {
             builder = new Builder(store, relationModelName)
-            .select(displayAttributeName)
+            .select(selectAttributes)
             .orderBy(`${displayAttributeName} ${_this.get('sorting')}`);
           }
 
@@ -1006,8 +1052,37 @@ export default FlexberryBaseComponent.extend({
         // and Ember won't change computed property.
 
         if (state !== 'selected') {
-          if (_this.get('displayValue')) {
-            _this.set('displayValue', _this._buildDisplayValue());
+          let displayValue = _this.get('displayValue');
+          if (displayValue) {
+            if (_this.get('autocompletePersistValue')) {
+              let builder = new Builder(store, relationModelName).select(displayAttributeName);
+
+              let autocompletePredicate = new SimplePredicate(displayAttributeName, 'eq', displayValue);
+              let resultPredicate = _this._conjuctPredicates(_this.get('lookupLimitPredicate'), autocompletePredicate);
+              if (resultPredicate) {
+                builder.where(resultPredicate);
+              }
+
+              run(() => {
+                store.query(relationModelName, builder.build()).then((records) => {
+                  let record = records.objectAt(0);
+                  if (isNone(record)) {
+                    _this.set('autocompletePersistValueCache', displayValue);
+                    _this.sendAction('remove', _this.get('removeData'));
+                  } else {
+                    _this.set('value', record);
+                    _this.get('currentController').send(_this.get('updateLookupAction'),
+                      {
+                        relationName: relationName,
+                        modelToLookup: relatedModel,
+                        newRelationValue: record
+                      });
+                  }
+                });
+              });
+            } else {
+              _this.set('displayValue', _this._buildDisplayValue());
+            }
           } else {
             _this.sendAction('remove', _this.get('removeData'));
           }
@@ -1130,10 +1205,15 @@ export default FlexberryBaseComponent.extend({
     @private
   */
   _buildDisplayValue() {
+    let autocompletePersistValueCache = this.get('autocompletePersistValueCache');
+    if (autocompletePersistValueCache) {
+      this.set('autocompletePersistValueCache', undefined);
+    }
+
     let selectedModel = this.get('value');
     let displayAttributeName = this.get('displayAttributeName');
     if (!selectedModel) {
-      return '';
+      return autocompletePersistValueCache ? autocompletePersistValueCache : '';
     }
 
     if (!displayAttributeName) {