+++ tests/acceptance/components/flexberry-objectlistview/filther/folv-filter-by-enther-click-test.js
@@ -53,7 +53,7 @@ executeTest('check filter by enter click', (store, assert, app) => {
             }
           }
 
-          assert.equal(filtherResult.length >= 1, true, 'Filtered list is not empty');
+          assert.equal(filtherResult.length >= 1, true, 'Filtered list is empty');
           assert.equal(successful, true, 'Filter successfully worked');
           done1();
         });