+++ tests/dummy/app/controllers/components-examples/flexberry-groupedit/ember-flexberry-dummy-suggestion-edit-groupedit-with-lookup-with-computed-atribute.js
@@ -1,5 +1,4 @@
-import { isNone } from '@ember/utils';
-import { observer } from '@ember/object';
+import { observer, computed } from '@ember/object';
 import BaseEditFormController from 'ember-flexberry/controllers/edit-form';
 import EditFormControllerOperationsIndicationMixin from 'ember-flexberry/mixins/edit-form-controller-operations-indication';
 import { StringPredicate } from 'ember-flexberry-data/query/predicate';
@@ -23,34 +22,44 @@ export default BaseEditFormController.extend(EditFormControllerOperationsIndicat
    */
   commentsEditRoute: 'ember-flexberry-dummy-comment-edit',
 
-  /**
-    Name of model.comments edit route.
-
-    @property commentsEditRoute
-    @type String
-    @default 'ember-flexberry-dummy-comment-edit'
-   */
   checkboxValue: false,
+
   fieldvalue: 'Vasya',
 
   lookupReadonly: observer('checkboxValue', function() {
-    if (!isNone(this.get('computedProperties.dynamicProperties.readonly'))) {
-      if (this.get('checkboxValue')) {
-        this.set('computedProperties.dynamicProperties.readonly', true);
-      } else {
-        this.set('computedProperties.dynamicProperties.readonly', false);
-      }
-    }
-
-    return this.get('checkboxValue');
+    this.set('lookupDynamicProperties.readonly', this.get('checkboxValue'));
   }),
 
   lookupLimitFunction: observer('fieldvalue', function() {
-    if (!isNone(this.get('computedProperties.dynamicProperties.lookupLimitPredicate'))) {
-      this.set('computedProperties.dynamicProperties.lookupLimitPredicate', new StringPredicate('name').contains(this.get('fieldvalue')));
+    this.set('lookupDynamicProperties.lookupLimitPredicate', new StringPredicate('name').contains(this.get('fieldvalue')));
+  }),
+
+  /**
+    An object with properties for the component `flexberry-lookup` in the component `flexberry-groupedit`.
+
+    @property lookupDynamicProperties
+    @type Object
+    @readOnly
+  */
+  lookupDynamicProperties: computed(function() {
+    let lookupLimitPredicate;
+    let fieldvalue = this.get('fieldvalue');
+    if (fieldvalue) {
+      lookupLimitPredicate = new StringPredicate('name').contains(fieldvalue);
     }
 
-  }),
+    return {
+      choose: 'showLookupDialog',
+      remove: 'removeLookupValue',
+      displayAttributeName: 'name',
+      required: true,
+      relationName: 'author',
+      projection: 'ApplicationUserL',
+      autocomplete: true,
+      readonly: this.get('checkboxValue'),
+      lookupLimitPredicate,
+    };
+  }).readOnly(),
 
   /**
     Method to get type and attributes of a component,
@@ -65,22 +74,10 @@ export default BaseEditFormController.extend(EditFormControllerOperationsIndicat
    */
   getCellComponent(attr, bindingPath, model) {
     let cellComponent = this._super(...arguments);
-    let limitFunction = new StringPredicate('name').contains('Vasya');
     if (attr.kind === 'belongsTo') {
       switch (`${model.modelName}+${bindingPath}`) {
         case 'ember-flexberry-dummy-vote+author':
-          cellComponent.componentProperties = {
-            choose: 'showLookupDialog',
-            remove: 'removeLookupValue',
-            displayAttributeName: 'name',
-            required: true,
-            relationName: 'author',
-            projection: 'ApplicationUserL',
-            autocomplete: true,
-            lookupLimitPredicate: limitFunction,
-            computedProperties: { thisController: this },
-            readonly: false,
-          };
+          cellComponent.componentProperties = this.get('lookupDynamicProperties');
           break;
 
         case 'ember-flexberry-dummy-comment+author':