+++ tests/dummy/app/controllers/components-examples/flexberry-lookup/settings-example.js
@@ -52,6 +52,15 @@ export default EditFormController.extend({
   autocomplete: false,
 
   /**
+    Flag indicates whether 'flexberry-lookup' component is in 'autocompletePersistValue' mode or not.
+
+    @property autocompletePersistValue
+    @type Boolean
+    @default false
+  */
+  autocompletePersistValue: false,
+
+  /**
     Flag indicates whether 'flexberry-lookup' component is in 'dropdown' mode or not.
 
     @property dropdown
@@ -119,6 +128,8 @@ export default EditFormController.extend({
       '  choose=(action "showLookupDialog")<br>' +
       '  remove=(action "removeLookupValue")<br>' +
       '  autocomplete=autocomplete<br>' +
+      '  autocompletePersistValue=autocompletePersistValue<br>' +
+      '  displayValue=model.lookupDisplayValue<br>' +
       '  dropdown=dropdown<br>' +
       '  chooseText=chooseText<br>' +
       '  removeText=removeText<br>' +
@@ -158,6 +169,12 @@ export default EditFormController.extend({
       settingType: 'boolean',
       settingDefaultValue: false,
       bindedControllerPropertieName: 'autocomplete'
+    });
+    componentSettingsMetadata.pushObject({
+      settingName: 'autocompletePersistValue',
+      settingType: 'boolean',
+      settingDefaultValue: false,
+      bindedControllerPropertieName: 'autocompletePersistValue'
     });
     componentSettingsMetadata.pushObject({
       settingName: 'dropdown',