+++ tests/dummy/app/models/ember-flexberry-dummy-application-user.js
@@ -57,4 +57,9 @@ Model.defineProjection('ApplicationUserL', 'ember-flexberry-dummy-application-us
   karma: attr('Karma')
 });
 
+// Projection for lookup example on window customization.
+Model.defineProjection('PreviewExampleView', 'ember-flexberry-dummy-application-user', {
+  name: attr('Name'),
+});
+
 export default Model;