+++ tests/integration/components/flexberry-lookup-test.js
@@ -41,7 +41,7 @@ moduleForComponent('flexberry-lookup', 'Integration | Component | flexberry look
 });
 
 test('component renders properly', function(assert) {
-  assert.expect(30);
+  assert.expect(31);
 
   this.render(hbs`{{#flexberry-lookup
   placeholder='(тестовое значение)'}}
@@ -51,6 +51,7 @@ test('component renders properly', function(assert) {
   let $component = this.$().children();
   let $lookupFluid = $component.children('.fluid');
   let $lookupInput = $lookupFluid.children('.lookup-field');
+  let $lookupButtonPreview = $lookupFluid.children('.ui-preview');
   let $lookupButtonChoose = $lookupFluid.children('.ui-change');
   let $lookupButtonClear = $lookupFluid.children('.ui-clear');
   let $lookupButtonClearIcon = $lookupButtonClear.children('.remove');
@@ -75,6 +76,9 @@ test('component renders properly', function(assert) {
   assert.strictEqual($lookupInput.hasClass('ember-view'), true, 'Component\'s title block has \'ember-view\' css-class');
   assert.strictEqual($lookupInput.hasClass('ember-text-field'), true, 'Component\'s title block has \'ember-text-field\' css-class');
   assert.equal($lookupInput.attr('placeholder'), '(тестовое значение)', 'Component\'s container has \'input\' css-class');
+
+  // Check <preview button>.
+  assert.strictEqual($lookupButtonPreview.length === 0, true, 'Component has inner title block');
 
   // Check <choose button>.
   assert.strictEqual($lookupButtonChoose.length === 1, true, 'Component has inner title block');
@@ -217,5 +221,114 @@ test('autocomplete doesn\'t send data-requests in readonly mode', function(asser
     $.ajax = originalAjaxMethod;
 
     asyncOperationsCompleted();
+  });
+});
+
+test('preview button renders properly', function(assert) {
+  assert.expect(11);
+
+  let store = app.__container__.lookup('service:store');
+
+  this.render(hbs`{{flexberry-lookup
+    value=model
+    relationName="parent"
+    projection="SuggestionTypeL"
+    displayAttributeName="name"
+    title="Parent"
+    showPreviewButton=true
+    previewFormRoute="ember-flexberry-dummy-suggestion-type-edit"
+  }}`);
+
+  // Retrieve component.
+  let $component = this.$().children();
+  let $lookupFluid = $component.children('.fluid');
+
+  assert.strictEqual($lookupFluid.children('.ui-preview').length === 0, true, 'Component has inner title block');
+
+  run(() => {
+    this.set('model', store.createRecord('ember-flexberry-dummy-suggestion-type', {
+      name: 'TestTypeName'
+    }));
+
+    let $lookupButtonPreview = $lookupFluid.children('.ui-preview');
+    let $lookupButtonPreviewIcon = $lookupButtonPreview.children('.eye');
+
+    assert.strictEqual($lookupButtonPreview.length === 1, true, 'Component has inner title block');
+    assert.strictEqual($lookupButtonPreview.prop('tagName'), 'BUTTON', 'Component\'s title block is a <button>');
+    assert.strictEqual($lookupButtonPreview.hasClass('ui'), true, 'Component\'s container has \'ui\' css-class');
+    assert.strictEqual($lookupButtonPreview.hasClass('ui-preview'), true, 'Component\'s container has \'ui-preview\' css-class');
+    assert.strictEqual($lookupButtonPreview.hasClass('button'), true, 'Component\'s container has \'button\' css-class');
+    assert.equal($lookupButtonPreview.attr('title'), 'Просмотр');
+
+    assert.strictEqual($lookupButtonPreviewIcon.length === 1, true, 'Component has inner title block');
+    assert.strictEqual($lookupButtonPreviewIcon.prop('tagName'), 'I', 'Component\'s title block is a <i>');
+    assert.strictEqual($lookupButtonPreviewIcon.hasClass('eye'), true, 'Component\'s container has \'eye\' css-class');
+    assert.strictEqual($lookupButtonPreviewIcon.hasClass('icon'), true, 'Component\'s container has \'icon\' css-class');
+  });
+});
+
+test('preview button view previewButtonClass and previewText properly', function(assert) {
+  assert.expect(3);
+
+  let store = app.__container__.lookup('service:store');
+
+  run(() => {
+    this.set('model', store.createRecord('ember-flexberry-dummy-suggestion-type', {
+      name: 'TestTypeName'
+    }));
+
+    this.set('previewButtonClass', 'previewButtonClassTest');
+    this.set('previewText', 'previewTextTest');
+
+    this.render(hbs`{{flexberry-lookup
+      value=model
+      relationName="parent"
+      projection="SuggestionTypeL"
+      displayAttributeName="name"
+      title="Parent"
+      showPreviewButton=true
+      previewFormRoute="ember-flexberry-dummy-suggestion-type-edit"
+      previewButtonClass=previewButtonClass
+      previewText=previewText
+    }}`);
+
+    // Retrieve component.
+    let $component = this.$().children();
+    let $lookupFluid = $component.children('.fluid');
+    let $lookupButtonPreview = $lookupFluid.children('.ui-preview');
+
+    assert.strictEqual($lookupButtonPreview.length === 1, true, 'Component has inner title block');
+    assert.strictEqual($lookupButtonPreview.hasClass('previewButtonClassTest'), true, 'Component\'s container has \'previewButtonClassTest\' css-class');
+    assert.equal($lookupButtonPreview.text().trim(), 'previewTextTest');
+  });
+});
+
+test('preview with readonly renders properly', function(assert) {
+  assert.expect(1);
+
+  let store = app.__container__.lookup('service:store');
+
+  run(() => {
+    this.set('model', store.createRecord('ember-flexberry-dummy-suggestion-type', {
+      name: 'TestTypeName'
+    }));
+
+    this.render(hbs`{{flexberry-lookup
+      value=model
+      relationName="parent"
+      projection="SuggestionTypeL"
+      displayAttributeName="name"
+      title="Parent"
+      showPreviewButton=true
+      previewFormRoute="ember-flexberry-dummy-suggestion-type-edit"
+      readonly=true
+    }}`);
+
+    // Retrieve component.
+    let $component = this.$().children();
+    let $lookupFluid = $component.children('.fluid');
+    let $lookupButtonPreview = $lookupFluid.children('.ui-preview');
+
+    assert.strictEqual($lookupButtonPreview.hasClass('disabled'), false, 'Component\'s container has not \'disabled\' css-class');
   });
 });